+++ spec/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160422195310) do
+ActiveRecord::Schema.define(version: 20160617224938) do
 
   create_table "alchemy_attachments", force: :cascade do |t|
     t.string   "name"
@@ -222,7 +222,6 @@ ActiveRecord::Schema.define(version: 20160422195310) do
     t.integer  "parent_id"
     t.integer  "depth"
     t.boolean  "visible",          default: false
-    t.boolean  "locked",           default: false
     t.integer  "locked_by"
     t.boolean  "restricted",       default: false
     t.boolean  "robot_index",      default: true
@@ -238,9 +237,11 @@ ActiveRecord::Schema.define(version: 20160422195310) do
     t.datetime "published_at"
     t.datetime "public_on"
     t.datetime "public_until"
+    t.datetime "locked_at"
   end
 
   add_index "alchemy_pages", ["language_id"], name: "index_pages_on_language_id"
+  add_index "alchemy_pages", ["locked_at", "locked_by"], name: "index_alchemy_pages_on_locked_at_and_locked_by"
   add_index "alchemy_pages", ["parent_id", "lft"], name: "index_pages_on_parent_id_and_lft"
   add_index "alchemy_pages", ["public_on", "public_until"], name: "index_alchemy_pages_on_public_on_and_public_until"
   add_index "alchemy_pages", ["urlname"], name: "index_pages_on_urlname"