+++ CHANGELOG.md
@@ -36,6 +36,7 @@ __Notable Changes__
 * Removes the `Language.get_default` method alias for `Language.default`
 * Move site select into pages and languages module to avoid confusion about curent site (#1067)
 * List pages from all sites in currently locked pages tabs and Dashboard widget (#1067)
+* The locked value on page is now a timestamp (`locked_at`), so we can order locked pages by (#1070)
 
 __Fixed Bugs__